Use Coordinates module instead of coord template on enwiki
Using the {{coord}} template has a high post-expand include size overhead and is causing pages to exceed the limit. This change can also apply to other wikis, but I'm not sure which other ones have either migrated to Lua or imported the template after the 2013 enwiki migration of the template.